Update core-ktx and material dependencies versions
Updated androidx.core:core-ktx from version 1.9.0 to 1.10.1 and com.google.android.material:material from version 1.8.0 to 1.9.0 in the project 'logstronaut'. These changes are part of an ongoing attempt to use the most stable and feature-rich versions in our project, to prevent potential bugs and to have access to new features.