Refactor logging method in MainActivity.kt
Moved the "Hello World" debug log to a separate function named testLog(). This was done to improve code readability and maintainability. This function now can be reused for different log messages in the future. Furthermore, android.util.Log was imported to show that this logging is connected with Android OS directly.