Replace LogItem with Loggable interface to improve flexibility
The 'LogItem' has been replaced with a 'Loggable' interface to permit more flexibility in handling logged items. As a direct alternative to the 'LogItem' class, the 'Loggable' interface maintains the same functionality but its abstract form will allow more customized implementations in the future. Consequently, 'LogstronautLogger' has been adjusted to operate with the new 'Loggable' interface instead of 'LogItem'.