Refactor tab functionality placement from Loggable to String
In this change, the function `Int.tabs()` was moved from `Loggable.kt` to `String.kt`. The `Int.tabs()` is a utility function that generates a string containing a specified number of tabs. This change was applied to improve the organization of the code, by putting this utility function with the rest of the string utility functions, this enhances the maintainability of the codebase.